+++ docs/PackageReleaseNotes/3.3.0/ReleaseNotes.md
@@ -1,10 +1,18 @@
 # FolkerKinzel.Strings 3.3.0
 ## Package Release Notes
 
+- Simulation of String.Create<TState>(int, TState, SpanAction<char, TState>) that works also for 
+older .NET Versions:
+```csharp
+string StringCreator.Create<TState>(int, TState, SpanAction<char, TState>);
+```
+
+.
+
 - New polyfills for .NET Framework 4.5 and .NET Standard 2.0:
 ```csharp
-int LastIndexOf(this ReadOnlySpan<char>, string?, StringComparison);
-int LastIndexOf(this ReadOnlySpan<char>, string?, int, int, StringComparison);
+int LastIndexOf(this ReadOnlySpan<char>, string, StringComparison);
+int LastIndexOf(this ReadOnlySpan<char>, string, int, int, StringComparison);
 
 System.Buffers.SpanAction<T, TArg>;
 ```